requirements_webapp.txt
streamlit==1.32.0
google-generativeai==0.3.2
torch==2.1.0+cpu
dgl==2.1.0
numpy==1.26.4
Pillow==10.2.0
matplotlib==3.8.3
pyverilog==1.3.0
python-dotenv==1.0.1
requests==2.32.3 